add netsrv pkg
